Template: vite-cf-DO-runner
Base template for History record of Mark Six of Hong Kong in recent draws:

8, 14, 16, 18, 26, 48, 44

13, 21, 33, 41, 44, 46, 43

15, 17, 19, 23, 24, 34, 33

1, 3, 24, 31, 39, 45, 7

5, 6, 18, 19, 30, 39, 7

3, 15, 17, 24, 32, 44, 20

2, 11, 32, 40, 43, 48, 12

5, 13, 17, 18, 31, 44, 2

1, 8, 9, 11, 18, 32, 13

4, 19, 24, 25, 26, 46, 39

6, 7, 27, 36, 39, 43, 1

4, 7, 15, 21, 45, 46, 24

19, 20, 26, 28, 39, 44, 4

10, 17, 22, 33, 40, 41, 31

7, 9, 16, 17, 33, 46, 49

18, 20, 28, 37, 38, 40, 41

10, 11, 28, 30, 37, 39, 15

2, 11, 13, 28, 38, 47, 7

2, 4, 10, 26, 28, 36, 23

19, 26, 33, 35, 36, 39, 5

1, 2, 17, 35, 37, 48, 8

6, 18, 29, 34, 37, 38, 39

4, 6, 26, 28, 34, 40, 25

Please summarize some rules that can predict 70% to 80% of numbers drawn.

The rules need not to be very strict and 100% correct.

Please generate Python code to draw 3 sets of high possibility with rules summarized above on these draws.